requirements.txt
python-telegram-bot==20.7
pymongo[srv]==4.6.1
dnspython
certifi
